refactor(sched): extract scheduler types into shared sched-types crate
Resolve the OperatorClass/BudgetResult/OperatorBudget duplication
between kernel and onnx-rt by extracting them into a new Layer 0
sched-types crate. Both kernel and onnx-rt now re-export from this
shared crate, eliminating the SonarCloud cpd.exclusions workaround
that was added in PR #69 (timer-hal-wcet-v1).

- New sched-types crate (no_std, alloc-free, no deps)
- kernel/src/sched/executor.rs: re-export instead of local defs
- onnx-rt/src/profile.rs: re-export instead of duplicate defs
- Remove sonar.cpd.exclusions for profile.rs
- Workspace grows from 20 to 21 crates

Pure refactor — all 3,200+ tests pass unchanged.
