+++ Justfile
@@ -15,7 +15,7 @@ build_std := "-Z build-std=core,compiler_builtins,alloc -Z build-std-features=co
 max_kernel_size_mb := "15"
 
 # Host-testable crates for module-level analysis
-host_crates := "smallaios-kernel smallaios-security smallaios-compute smallaios-onnx-rt smallaios-ipc smallaios-net smallaios-posix smallaios-container smallaios-bus smallaios-peripheral smallaios-usb smallaios-sdr smallaios-bench"
+host_crates := "smallaios-kernel smallaios-security smallaios-compute smallaios-sched-types smallaios-onnx-rt smallaios-ipc smallaios-net smallaios-posix smallaios-container smallaios-bus smallaios-peripheral smallaios-usb smallaios-sdr smallaios-bench"
 
 # === Container Mode (Library OS) ===
