+++ kernel/src/sched/executor.rs
@@ -138,84 +138,7 @@ impl RunQueue {
     }
 }
 
-/// Per-operator time budget configuration.
-#[derive(Debug, Clone, Copy)]
-pub struct OperatorBudget {
-    /// Budget for elementwise ops (Relu, Add, etc.) in microseconds.
-    pub elementwise_us: u64,
-    /// Budget for reduction ops (Softmax, ReduceMean) in microseconds.
-    pub reduction_us: u64,
-    /// Budget for GEMM ops (MatMul, Conv) in microseconds.
-    pub gemm_us: u64,
-    /// Budget for attention ops in microseconds.
-    pub attention_us: u64,
-    /// Budget for GPU kernels in microseconds.
-    pub gpu_kernel_us: u64,
-    /// Soft limit multiplier (warning at 1x, soft at this, hard at hard_multiplier).
-    pub soft_multiplier: u64,
-    /// Hard limit multiplier (abort inference at this multiple of budget).
-    pub hard_multiplier: u64,
-}
-
-impl OperatorBudget {
-    /// Default budgets per the spec.
-    pub const DEFAULT: Self = Self {
-        elementwise_us: 1_000,    // 1 ms
-        reduction_us: 10_000,     // 10 ms
-        gemm_us: 100_000,         // 100 ms
-        attention_us: 500_000,    // 500 ms
-        gpu_kernel_us: 1_000_000, // 1000 ms
-        soft_multiplier: 2,
-        hard_multiplier: 10,
-    };
-}
-
-/// Operator type for budget lookup.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-#[repr(u8)]
-pub enum OperatorClass {
-    Elementwise = 0,
-    Reduction = 1,
-    Gemm = 2,
-    Attention = 3,
-    GpuKernel = 4,
-}
-
-/// Result of checking an operator's execution time against its budget.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum BudgetResult {
-    /// Within budget — continue normally.
-    Ok,
-    /// Exceeded 1x budget — log warning.
-    Warning,
-    /// Exceeded soft limit — emit metric, flag for profiling.
-    SoftLimit,
-    /// Exceeded hard limit — abort inference.
-    HardLimit,
-}
-
-impl OperatorBudget {
-    /// Check an operator's actual execution time against its budget.
-    pub fn check(&self, op_class: OperatorClass, actual_us: u64) -> BudgetResult {
-        let budget = match op_class {
-            OperatorClass::Elementwise => self.elementwise_us,
-            OperatorClass::Reduction => self.reduction_us,
-            OperatorClass::Gemm => self.gemm_us,
-            OperatorClass::Attention => self.attention_us,
-            OperatorClass::GpuKernel => self.gpu_kernel_us,
-        };
-
-        if actual_us > budget * self.hard_multiplier {
-            BudgetResult::HardLimit
-        } else if actual_us > budget * self.soft_multiplier {
-            BudgetResult::SoftLimit
-        } else if actual_us > budget {
-            BudgetResult::Warning
-        } else {
-            BudgetResult::Ok
-        }
-    }
-}
+pub use smallaios_sched_types::{BudgetResult, OperatorBudget, OperatorClass};
 
 /// Watchdog state managed by the executor.
 pub struct WatchdogState {
