+++ sonar-project.properties
@@ -4,11 +4,3 @@ sonar.sources=kernel/src,security/src,onnx-rt/src,ipc/src,net/src,posix/src,cont
 sonar.exclusions=arch/**,target/**
 sonar.language=rust
 sonar.coverageReportPaths=sonar-coverage.xml
-
-# Duplication exclusions — onnx-rt/src/profile.rs intentionally mirrors
-# kernel/src/sched/executor.rs types (OperatorClass, BudgetResult,
-# OperatorBudget). The architecture layering prevents onnx-rt (Layer 1)
-# from depending on kernel (Layer 0). A future sched-types crate at
-# Layer 0 will unify these types. See openspec/changes/archive/
-# 2026-04-10-timer-hal-wcet-v1/design.md D1 for rationale.
-sonar.cpd.exclusions=onnx-rt/src/profile.rs
